merapikan tampilan create dan juga edit menjadi satu baris saja,menghapus setiap label pada create dan edit items customer dan juga itemscustomer,kemudian saya membuat notif ketika master data ingin dihapus tidak bisa karena data dari master data pada child nya sedang digunakan,memperbaiki fitur search pada invoice yang dimana ketika user search value dari user tadi di search masih ada dan ketika tombol reset ditekan maka akan mereset semua yang tadi di inputkan oleh user

+++ controllers/itemsController.php
@@ -85,11 +85,21 @@ else if (isset($_GET['detail_item'])) {
 // Hapus item
 else if (isset($_GET['delete_item'])) {
     $id = $_GET['delete_item'];
-    $itemModel->delete($id);
-    $_SESSION['alert_delete'] = [
-        'type' => 'success',
-        'message' => 'Item berhasil dihapus!'
-    ];
+
+    if($itemModel->isUseId($id)){
+        $_SESSION['alert_delete'] = [
+            'type' => 'danger',
+            'message' => 'Item tidak dapat dihapus karena masih digunakan di invoice items, atau items customers.'
+        ];
+    }
+    else{
+        $itemModel->delete($id);
+        $_SESSION['alert_delete'] = [
+            'type' => 'success',
+            'message' => 'Item berhasil dihapus!'
+        ];
+    }
+    
     header('Location:' . BASE_URL . 'pages/dataItems.php');
     exit();
 }
